+++ packages/wagtail/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @baseapp-frontend/wagtail
 
+## 1.0.62
+
+### Patch Changes
+
+- Updated dependencies
+  - @baseapp-frontend/graphql@1.4.3
+
+## 1.0.61
+
+### Patch Changes
+
+- Updated dependencies
+  - @baseapp-frontend/utils@4.0.9
+  - @baseapp-frontend/graphql@1.4.2
+  - @baseapp-frontend/design-system@1.2.13
+
 ## 1.0.60
 
 ### Patch Changes